draft: ledgerTx initial

+++ lib/omc/ledgers/ledger_acc.ex
@@ -5,7 +5,7 @@ defmodule Omc.Ledgers.LedgerAcc do
   schema "ledger_accs" do
     field :ledger_id, :id
     field :server_acc_id, :id
-    field :price_per_day, :decimal
+    field :price_per_day, :integer
     field :activated_at, :naive_datetime
     field :deactivated_at, :naive_datetime
   end
